Source tweaks to make comparing changes between 1.3 and 1.4 easier

+++ MTA10/mods/deathmatch/logic/CPacketHandler.cpp
@@ -4415,36 +4415,33 @@ void CPacketHandler::Packet_ResourceStart ( NetBitStreamInterface& bitStream )
                         }
 
                         // Is it a valid downloadable resource?
-                        if ( pDownloadableResource )
+                        if ( pDownloadableResource && pDownloadableResource->IsAutoDownload() )
                         {
-                            if ( pDownloadableResource->IsAutoDownload() )
+                            // Does the Client and Server checksum Match?
+                            if ( !pDownloadableResource->DoesClientAndServerChecksumMatch () )
                             {
-                                // Does the Client and Server checksum Match?
-                                if ( !pDownloadableResource->DoesClientAndServerChecksumMatch () )
+                                // Make sure the directory exists
+                                const char* szTempName = pDownloadableResource->GetName ();
+                                if ( szTempName )
                                 {
-                                    // Make sure the directory exists
-                                    const char* szTempName = pDownloadableResource->GetName ();
-                                    if ( szTempName )
-                                    {
-                                        // Make sure its directory exists
-                                        MakeSureDirExists ( szTempName );
-                                    }
+                                    // Make sure its directory exists
+                                    MakeSureDirExists ( szTempName );
+                                }
 
-                                    // Combine the HTTP Download URL, the Resource Name and the Resource File
-                                    SString strHTTPDownloadURLFull ( "%s/%s/%s", g_pClientGame->m_strHTTPDownloadURL.c_str (), pResource->GetName (), pDownloadableResource->GetShortName () );
+                                // Combine the HTTP Download URL, the Resource Name and the Resource File
+                                SString strHTTPDownloadURLFull ( "%s/%s/%s", g_pClientGame->m_strHTTPDownloadURL.c_str (), pResource->GetName (), pDownloadableResource->GetShortName () );
 
-                                    // Delete the file that already exists
-                                    unlink ( pDownloadableResource->GetName () );
+                                // Delete the file that already exists
+                                unlink ( pDownloadableResource->GetName () );
 
-                                    // Queue the file to be downloaded
-                                    bool bAddedFile = pHTTP->QueueFile ( strHTTPDownloadURLFull, pDownloadableResource->GetName (), dChunkDataSize, NULL, 0, false, NULL, NULL, g_pClientGame->IsLocalGame () );
+                                // Queue the file to be downloaded
+                                bool bAddedFile = pHTTP->QueueFile ( strHTTPDownloadURLFull, pDownloadableResource->GetName (), dChunkDataSize, NULL, 0, false, NULL, NULL, g_pClientGame->IsLocalGame () );
 
-                                    // If the file was successfully queued, increment the resources to be downloaded
-                                    usResourcesToBeDownloaded++;
-                                    if ( bAddedFile )
-                                        g_pClientGame->m_pTransferBox->AddToTotalSize ( dChunkDataSize );
-                                }                       
-                            }
+                                // If the file was successfully queued, increment the resources to be downloaded
+                                usResourcesToBeDownloaded++;
+                                if ( bAddedFile )
+                                    g_pClientGame->m_pTransferBox->AddToTotalSize ( dChunkDataSize );
+                            }                       
                         }
                     }
                 }
